Functions of matrices and linear maps

Functions of matrices and linear maps

Exponential function

KrylovKit.exponentiate — Function.

function exponentiate(A, t::Number, v; kwargs...)
function exponentiate(A, t::Number, v, algorithm)

Compute $exp(t*A) v$, where A is a general linear map, i.e. a AbstractMatrix or just a general function or callable object and v is of any Julia type with vector like behavior.

source