Available algorithms

Available algorithms

Orthogonalization algorithms

KrylovKit.ClassicalGramSchmidt — Type.

ClassicalGramSchmidt()

Represents the classical Gram Schmidt algorithm for orthogonalizing different vectors, typically not an optimal choice.

source

KrylovKit.ModifiedGramSchmidt — Type.

ModifiedGramSchmidt()

Represents the modified Gram Schmidt algorithm for orthogonalizing different vectors, typically a reasonable choice for linear systems but not for eigenvalue solvers with a large Krylov dimension.

source

KrylovKit.ClassicalGramSchmidt2 — Type.

ClassicalGramSchmidt2()

Represents the classical Gram Schmidt algorithm with a second reorthogonalization step always taking place.

source

KrylovKit.ModifiedGramSchmidt2 — Type.

ModifiedGramSchmidt2()

Represents the modified Gram Schmidt algorithm with a second reorthogonalization step always taking place.

source

KrylovKit.ClassicalGramSchmidtIR — Type.

ClassicalGramSchmidtIR(η::Real)

Represents the classical Gram Schmidt algorithm with zero or more reorthogonalization steps being applied untill the norm of the vector after an orthogonalization step has not decreased by a factor smaller than η with respect to the norm before the step.

source

KrylovKit.ModifiedGramSchmidtIR — Type.

ModifiedGramSchmidtIR(η::Real)

Represents the modified Gram Schmidt algorithm with zero or more reorthogonalization steps being applied untill the norm of the vector after an orthogonalization step has not decreased by a factor smaller than η with respect to the norm before the step.

source

General Krylov algorithms

KrylovKit.Lanczos — Type.

Lanczos(orth::Orthogonalizer = KrylovDefaults.orth; krylovdim = KrylovDefaults.krylovdim,
    maxiter::Int = KrylovDefaults.maxiter, tol = KrylovDefaults.tol)

Represents the Lanczos algorithm for building the Krylov subspace; assumes the linear operator is real symmetric or complex Hermitian. Can be used in eigsolve and exponentiate. The corresponding algorithms will build a Krylov subspace of size at most krylovdim, which will be repeated at most maxiter times and will stop when the norm of the residual of the Lanczos factorization is smaller than tol. The orthogonalizer orth will be used to orthogonalize the different Krylov vectors.

Use Arnoldi for non-symmetric or non-Hermitian linear operators.

See also: factorize, eigsolve, exponentiate, Arnoldi, Orthogonalizer

source

KrylovKit.Arnoldi — Type.

Arnoldi(orth::Orthogonalizer = KrylovDefaults.orth; krylovdim = KrylovDefaults.krylovdim,
    maxiter::Int = KrylovDefaults.maxiter, tol = KrylovDefaults.tol)

Represents the Arnoldi algorithm for building the Krylov subspace for a general matrix or linear operator. Can be used in eigsolve and exponentiate. The corresponding algorithms will build a Krylov subspace of size at most krylovdim, which will be repeated at most maxiter times and will stop when the norm of the residual of the Arnoldi factorization is smaller than tol. The orthogonalizer orth will be used to orthogonalize the different Krylov vectors.

Use Lanczos for real symmetric or complex Hermitian linear operators.

See also: eigsolve, exponentiate, Lanczos, Orthogonalizer

source

Specific algorithms for linear systems

CG
MINRES
GMRES
BiCG
BiCGStab

Specific algorithms for sigular values

KrylovKit.GKL — Type.

GKL(orth::Orthogonalizer = KrylovDefaults.orth; krylovdim = KrylovDefaults.krylovdim,
    maxiter::Int = KrylovDefaults.maxiter, tol = KrylovDefaults.tol)

Represents the Golub-Kahan-Lanczos bidiagonalization algorithm for sequentially building a Krylov-like factorization of a genereal matrix or linear operator with a bidiagonal reduced matrix. Can be used in svdsolve. The corresponding algorithm builds a Krylov subspace of size at most krylovdim, which will be repeated at most maxiter times and will stop when the norm of the residual of the Arnoldi factorization is smaller than tol. The orthogonalizer orth will be used to orthogonalize the different Krylov vectors.

See also: svdsolve, Orthogonalizer

source

Default values

KrylovKit.KrylovDefaults — Module.

module KrylovDefaults
    const orth = KrylovKit.ModifiedGramSchmidtIR(1/sqrt(2))
    const krylovdim = 30
    const maxiter = 100
    const tol = 1e-12
end

A module listing the default values for the typical parameters in Krylov based algorithms:

orth: the orthogonalization routine used to orthogonalize the Krylov basis in the Lanczos or Arnoldi iteration

krylovdim: the maximal dimension of the Krylov subspace that will be constructed

maxiter: the maximal number of outer iterations, i.e. the maximum number of times the Krylov subspace may be rebuilt

tol: the tolerance to which the problem must be solved, based on a suitable error measure, e.g. the norm of some residual. !!! warning The default value of tol is a Float64 value, if you solve problems in Float32 or ComplexF32 arithmetic, you should always specify a new tol as the default value will not be attainable.

source